fix repo package by move unrelated interfaces
