+++ src/main/kotlin/logic/usecase/KetoDietMealHelperUseCase.kt
@@ -1,8 +1,8 @@
 package org.example.logic.usecase
 
-import org.example.logic.repository.CalculatePercentage
+import org.example.logic.CalculatePercentage
 import org.example.logic.repository.MealRepository
-import org.example.logic.repository.ReactionProvider
+import org.example.logic.ReactionProvider
 import org.example.model.Meal
 
 class KetoDietMealHelperUseCase(
